move `avoid-setjmp-longjmp` from `bugprone` to `modernize`
