+++ clang-tools-extra/docs/clang-tidy/checks/cert/err52-cpp.rst
@@ -4,5 +4,9 @@ cert-err52-cpp
 ==============
 
 The cert-err52-cpp check is an alias, please see
-`bugprone-avoid-setjmp-longjmp <../bugprone/avoid-setjmp-longjmp.html>`_
+`modernize-avoid-setjmp-longjmp <../modernize/avoid-setjmp-longjmp.html>`_
 for more information.
+
+This check corresponds to the CERT C++ Coding Standard rule
+`ERR52-CPP. Do not use setjmp() or longjmp()
+<https://www.securecoding.cert.org/confluence/pages/viewpage.action?pageId=1834>`_.
